SAPPORO-service/requirements.dev.txt
bagit==1.7.0
CacheControl==0.11.7
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cwltool==1.0.20181217162649
decorator==4.3.0
Flask==1.0.2
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10
lockfile==0.12.2
lxml==4.2.5
MarkupSafe==1.1.0
mistune==0.8.4
mypy-extensions==0.4.1
networkx==2.2
prov==1.5.1
psutil==5.4.8
pyparsing==2.3.0
python-dateutil==2.7.5
python-dotenv==0.10.1
PyYAML==3.13
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.21.0
ruamel.yaml==0.15.77
scandir==1.9.0
schema-salad==3.0.20181206233650
shellescape==3.4.1
six==1.12.0
typing-extensions==3.6.6
urllib3==1.24.1
Werkzeug==0.14.1
